Allow updating binder when journal files are changed.
This is taken out of the BINDER function.
The new function BINDER-REFRESH-JOURNALS-IF-NEEDED is exported.

Ideally this new function should not exist and the BINDER function
will do instead.   However there is a design issue in BINDER which
makes this hard to realize.

BINDER will create a new binder object, and will transfer the journals
over to the new object.  However, the root account of all the
transactions is not updated.